Implement "chunked" writes
This implements writing only what is available in the transmit buffer.
This allows a severe reduction in blocking time - prior behavior would
continue blocking until the remainder could fit within the transmit buffer.

Requires a transmit done interrupt to be defined, otherwise it will
fall back to blocking for the entirety of the write.