Nonblocking receive: convert poll() to state machine
To implement nonblocking receive operation, convert poll()
to a state machine with five states:

 - STATE_IDLE
 - STATE_FRAME_RX
 - STATE_PARSE
 - STATE_BUILD_REPLY
 - STATE_TRANSMIT

Respect character and frame timeouts where necessary.

Also check if the Tx buffer is empty (or, fully available)
when checking if we're currently still transmitting -
on AVR, we will not have the IRQ to signal to us that
transmission has finished. This should prevent the board
from resetting or changing settings prior to finishing its
reply.